Boss moves: lock on fix (again); -Stab can slow down enemies; -Doppel attack stop on banish; -Doppel camera fix (for real); -Doppel not breaking stab camera anymore; -Pl0300 controller now loads pfb without PrefabManager, boss costume can be selected; -Fixed bug when game desides to change photomode camera to pl0800 pos during air raid; -Boss moves can be used with a few pl0800 in theory; -Pl0300 now realods in every bp stage like default players; EmSwapper: less shit code, less bp crash, less bp black screen, STONKS; -Driendly & bob vergil now uses non-EnemyPrefabManager prefabs, costume/ex costume can be selected; -Keybinds for enemy spawn added; PlayerTraker.on_pl_add() now passes PlayerManager; 3JCE: unused detours removed

+++ src/mods/Pl0300Controller.cpp
@@ -3,18 +3,19 @@
 
 void PlCntr::Pl0300Cntr::Pl0300Controller::destroy_game_obj()
 {
+	destroy_all_related_shells();
+	if (_isDoppelDestroyRequested)
+		return;
 	if (get_pl() != 0 && !is_doppel())
 	{
-		destroy_all_related_shells();
 		if (is_in_players_list())
 		{
-			sdk::call_object_func_easy<void*>(get_pl_manager(), "removePlayer(app.Player, System.Boolean)", (REManagedObject*)get_pl(), false);
+			sdk::call_object_func_easy<void*>(sdk::get_managed_singleton<REManagedObject>("app.PlayerManager"), "removePlayer(app.Player, System.Boolean)", get_pl(), false);
 		}
 		if (auto dp = _doppel.lock(); dp != nullptr)
 		{
 			destroy_doppel();
 		}
-
 		PlCntr::PlController::destroy_game_obj();
 	}
 }
@@ -53,41 +54,44 @@ PlCntr::Pl0300Cntr::Pl0300Controller::Pl0300Controller(uintptr_t pl0300, PlCntr:
 		throw std::exception("pl0300 is null");
 	_pl0300Type = type;
 
-	if(_pl0300SetDtMethod == nullptr)
-		_pl0300SetDtMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setDevilTrigger(app.player.PlayerBase.DevilTrigger, System.Boolean)");
+	if(_isStaticInitRequested)
+	{
+		if (_pl0300SetDtMethod == nullptr)
+			_pl0300SetDtMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setDevilTrigger(app.player.PlayerBase.DevilTrigger, System.Boolean)");
 
-	if(_pl0300GenDoppelMethod == nullptr)
-		_pl0300GenDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "generateDoppleGanger(System.Boolean)");
+		if (_pl0300GenDoppelMethod == nullptr)
+			_pl0300GenDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "generateDoppleGanger(System.Boolean)");
 
-	if (_pl0300DestroyDoppelMethod == nullptr)
-		_pl0300DestroyDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "destroyDoppleGanger()");
+		if (_pl0300DestroyDoppelMethod == nullptr)
+			_pl0300DestroyDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "destroyDoppleGanger()");
 
-	if (_pl0300DoStartMethod == nullptr)
-		_pl0300DoStartMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "doStart()");
+		if (_pl0300DoStartMethod == nullptr)
+			_pl0300DoStartMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "doStart()");
 
-	if (_pl0300RequestDestroyDoppelMethod == nullptr)
-		_pl0300RequestDestroyDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "requestDestroyDoppelGanger()");
+		if (_pl0300RequestDestroyDoppelMethod == nullptr)
+			_pl0300RequestDestroyDoppelMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "requestDestroyDoppelGanger()");
 
-	if (_baseBehaviorStartMethod == nullptr)
-		_baseBehaviorStartMethod = sdk::find_method_definition("app.BaseBehavior", "start()");
+		if (_baseBehaviorStartMethod == nullptr)
+			_baseBehaviorStartMethod = sdk::find_method_definition("app.BaseBehavior", "start()");
 
-	if (_pl0300startAirRaidMethod == nullptr)
-		_pl0300startAirRaidMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "startAirRaid()");
+		if (_pl0300startAirRaidMethod == nullptr)
+			_pl0300startAirRaidMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "startAirRaid()");
 
-	if(_pl0300SetActionMethod == nullptr)
-		_pl0300SetActionMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setAction(System.String, System.UInt32, System.Single, System.Single, "
-			"via.motion.InterpolationMode, via.motion.InterpolationCurve, System.Boolean, System.Boolean, System.Boolean, app.GameModel.ActionPriority)");
+		if (_pl0300SetActionMethod == nullptr)
+			_pl0300SetActionMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setAction(System.String, System.UInt32, System.Single, System.Single, "
+				"via.motion.InterpolationMode, via.motion.InterpolationCurve, System.Boolean, System.Boolean, System.Boolean, app.GameModel.ActionPriority)");
 
-	if (_pl0300SetActionFromThinkMethod == nullptr)
-		_pl0300SetActionFromThinkMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setActionFromThink(System.String, System.UInt32)");
+		if (_pl0300SetActionFromThinkMethod == nullptr)
+			_pl0300SetActionFromThinkMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "setActionFromThink(System.String, System.UInt32)");
 
-	if (_pl0300UpdateEmTeleportMethod == nullptr)
-		_pl0300UpdateEmTeleportMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "updateEnemyTeleport(System.Boolean)");
+		if (_pl0300UpdateEmTeleportMethod == nullptr)
+			_pl0300UpdateEmTeleportMethod = sdk::find_method_definition("app.player.pl0300.PlayerVergil", "updateEnemyTeleport(System.Boolean)");
 
-	//_baseBehaviorStartMethod->call(sdk::get_thread_context(), _pl0300);
-	//pl0300_do_start();
-	if (_pl0300Manager == nullptr)
-		_pl0300Manager = static_cast<Pl0300ControllerManager*>(g_framework->get_mods()->get_mod("Pl0300ControllerManager"));
+		if (_pl0300Manager == nullptr)
+			_pl0300Manager = static_cast<Pl0300ControllerManager*>(g_framework->get_mods()->get_mod("Pl0300ControllerManager"));
+
+		_isStaticInitRequested = false;
+	}
 }
 
 void PlCntr::Pl0300Cntr::Pl0300Controller::destroy_doppel()
@@ -97,10 +101,11 @@ void PlCntr::Pl0300Cntr::Pl0300Controller::destroy_doppel()
 		return;
 	auto groupTmp = get_char_group();
 	change_character_group(PlCntr::CharGroup::Enemy);
+	/*if (_pl0300RequestDestroyDoppelMethod != nullptr)
+		_pl0300RequestDestroyDoppelMethod->call(sdk::get_thread_context(), _doppel.lock()->get_pl());*/
 	if (_pl0300DestroyDoppelMethod != nullptr)
 		_pl0300DestroyDoppelMethod->call(sdk::get_thread_context(), get_pl());
-	if (_pl0300RequestDestroyDoppelMethod != nullptr)
-		_pl0300RequestDestroyDoppelMethod->call(sdk::get_thread_context(), get_pl());
+	_doppel.lock()->_isDoppelDestroyRequested = true;
 	_pl0300Manager->remove_doppelganger(_doppel.lock().get());
 	change_character_group(groupTmp);
 }
